backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
python-dotenv
faiss-cpu
openai
docling 
docling-core 
pypdf
azure-search-documents
azure-identity
python-multipart

